requirement.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.6.1
jsonpath==0.82
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-assume==2.2.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
